+++ packages/pwa-kit-create-app/CHANGELOG.md
@@ -1,4 +1,6 @@
 ## v3.8.0-dev (Aug 8, 2024)
+
+- Removed OCAPISessionURL prop from provider template. [#2090](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2090)
 - Update ssr.js templates to include new feature flag to encode non ASCII HTTP headers [#2048](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2048)
 - Replace getAppOrigin with useOrigin to have a better support for an app origin building. [#2050](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2050)